requirements.txt
opencv-python
matplotlib
numpy
easyocr
mysql-connector-python
flask
requests